requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
gunicorn
langchain-google-genai
python-dotenv
langchain
langchain-community
langchainhub
google-generativeai
requests
qdrant-client
